Move tests and change yarn clean regex
