+++ client/package.json
@@ -13,8 +13,8 @@
     "test": "jest",
     "dist": "yarn build && electron-builder",
     "pack": "yarn build --dir && electron-builder",
-    "clean": "yarn prettier --write \"@(src|test)/**/*.@(js?(x)|scss)\" *.js ",
-    "clean-check": "yarn prettier --check \"@(src|test)/**/*.@(js?(x)|scss)\" *.js "
+    "clean": "yarn prettier --write \"src/**/*.@(js?(x)|scss)\" *.js ",
+    "clean-check": "yarn prettier --check \"src/**/*.@(js?(x)|scss)\" *.js "
   },
   "build": {
     "appId": "io.seanbailey.kauri",
